3548: AC.cpp (#1465) - AC,27.27%,32.95%
刚刚估计是负数long long的int变成正数了

+++ Codes/3548-equal-sum-grid-partition-ii.cpp
@@ -2,7 +2,7 @@
  * @Author: LetMeFly
  * @Date: 2026-03-26 21:49:18
  * @LastEditors: LetMeFly.xyz
- * @LastEditTime: 2026-03-26 22:26:44
+ * @LastEditTime: 2026-03-26 22:31:48
  */
 #ifdef _DEBUG
 #include "_[1,2]toVector.h"
@@ -37,7 +37,7 @@ private:
     // now = (all + x) / 2
     // x = now * 2 - all
     bool ok(vector<vector<int>>& grid, ll all) {
-        unordered_set<int> visited;
+        unordered_set<ll> visited;
         ll now = 0;
         for (int i = 0; i < grid.size(); i++) {
             for (int& t : grid[i]) {
